+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/dao/ClienteDAO.java
@@ -1,10 +1,10 @@
-package dao;
+package edu.uclm.esi.iso.ISO2023.dao;
 
 import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import entities.Administrador;
-import entities.Cliente;
+import edu.uclm.esi.iso.ISO2023.entities.Administrador;
+import edu.uclm.esi.iso.ISO2023.entities.Cliente;
 
 
 public interface ClienteDAO extends MongoRepository<Cliente,String>{
