+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/dao/TokenDAO.java
@@ -1,12 +1,12 @@
-package dao;
+package edu.uclm.esi.iso.ISO2023.dao;
 
 import java.util.Optional; 
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import entities.Administrador;
-import entities.Token;
-import entities.User;
+import edu.uclm.esi.iso.ISO2023.entities.Administrador;
+import edu.uclm.esi.iso.ISO2023.entities.Token;
+import edu.uclm.esi.iso.ISO2023.entities.User;
 
 
 public interface TokenDAO extends MongoRepository<Token,String>{
