+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/dao/VehiculoDAO.java
@@ -1,11 +1,11 @@
-package dao;
+package edu.uclm.esi.iso.ISO2023.dao;
 
 import java.util.Optional; 
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import entities.Administrador;
-import entities.Vehiculo;
+import edu.uclm.esi.iso.ISO2023.entities.Administrador;
+import edu.uclm.esi.iso.ISO2023.entities.Vehiculo;
 
 
 public interface VehiculoDAO extends MongoRepository<Vehiculo,String>{
