+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/entities/Administrador.java
@@ -1,6 +1,8 @@
-package entities;
+package edu.uclm.esi.iso.ISO2023.entities;
 
 import org.springframework.data.annotation.Id;
+
+import edu.uclm.esi.iso.ISO2023.entities.User;
 
 public class Administrador extends User{
