+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/entities/Cliente.java
@@ -1,5 +1,6 @@
-package entities;
+package edu.uclm.esi.iso.ISO2023.entities;
 
+import edu.uclm.esi.iso.ISO2023.entities.User;
 
 public class Cliente extends User{
