+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/entities/Patinete.java
@@ -1,6 +1,8 @@
-package entities;
+package edu.uclm.esi.iso.ISO2023.entities;
 
-import org.springframework.data.annotation.Id; 
+import org.springframework.data.annotation.Id;
+
+import edu.uclm.esi.iso.ISO2023.entities.Vehiculo; 
 
 public class Patinete extends Vehiculo{
