+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/entities/Token.java
@@ -1,6 +1,8 @@
-package entities;
+package edu.uclm.esi.iso.ISO2023.entities;
 
 import java.util.UUID;
+
+import edu.uclm.esi.iso.ISO2023.entities.User;
 
 public class Token {
 	private String id;
