+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/exceptions/contraseñaIncorrecta.java
@@ -1,4 +1,4 @@
-package exceptions;
+package edu.uclm.esi.iso.ISO2023.exceptions;
 /*Excepción personalizada que representa la invalidez de las contraseñas al no ser correcta*/
 
 public class contraseñaIncorrecta extends Exception{
